working implementation of adding a person to the canvas
notice: compared to the last commit, the behavior has changed.
rather than click-move-click, the behavior is press-drag-unpress.
i.e., you have to drag the person onto the canvas

this was motivated by the fact that mouseMoved events are lost (not reported)
when over the JLayeredPane's child components,
while mouseDragged events are retained.
(why? who knows...)